refactor and add form component
