+++ src/App.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import CreditCardForm from "./Components/CreditCardForm";
+
+function App() {
+  return <CreditCardForm />;
+}
+
+export default App;
